pyproject.toml
[tool.poetry]
name = "chatpdf"
version = "0.1.0"
description = ""
authors = ["Junichi Kakisako <kilo.sw@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
llama-index = "^0.5.10"
# pdfminer-six = "^20221105"
click = "^8.1.3"
langchain = "^0.0.134"
python-dotenv = "^1.0.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[tool.poetry.scripts]
chatpdf = "chatpdf.scripts.chatpdf:cli"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
